Add L2 table

+++ arch/arm64/mm/pgtable.h
@@ -11,7 +11,7 @@
 #define PTE_TABLE ((uint64_t)1 << 1)
 #define PTE_AF ((uint64_t)1 << 10)
 #define PTE_SH_INNER ((uint64_t)3 << 8)
-#define PTE_ATTRINDX(idx) ((idx) << 2)
+#define PTE_ATTRINDX(idx) ((uint64_t)(idx) << 2)
 #define PTE_PXN ((uint64_t)1 << 53)
 #define PTE_UXN ((uint64_t)1 << 54)
